rsx: Refactor out complex present code into separate files
- Also restructures present code to have image lookup in a separate
re-usable function.
